requirements.txt
crewai>=0.28.0
crewai-tools>=0.3.0
python-dotenv>=1.0.0
openai>=1.3.0
